solved problme of migrations and resolved #66

+++ database/migrations/2024_11_08_143429_create_product_recipes_table.php
@@ -4,23 +4,24 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('product_recipes', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('recipe_id')->constrained()->onDelete('cascade');
-            $table->foreignId('product_id')->constrained()->onDelete('cascade');
-            $table->foreignId('product_variant_id')->nullable()->constrained()->onDelete('cascade');
-            $table->boolean('is_default')->default(false);
-            $table->timestamps();
+        if (!Schema::hasTable('product_recipes')) {
+            Schema::create('product_recipes', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('recipe_id')->constrained()->onDelete('cascade');
+                $table->foreignId('product_id')->constrained()->onDelete('cascade');
+                $table->foreignId('product_variant_id')->nullable()->constrained()->onDelete('cascade');
+                $table->boolean('is_default')->default(false);
+                $table->timestamps();
 
-            $table->unique(['recipe_id', 'product_id', 'product_variant_id']);
-        });
+                $table->unique(['recipe_id', 'product_id', 'product_variant_id']);
+            });
+        }
     }
 
     /**
